refactor: SiteSearch component, SiteWrapper with data, and Layout for the Page
